Fix FreememHealthCheck Int overflow and PulsarHealthCheck no-op connectivity check (#142)
FreememHealthCheck: factory methods mb() and gb() used Int arithmetic that
silently overflowed to negative values for inputs ≥ 2GB, causing the threshold
to always compare as less than available memory and the check to always pass.
Fixed by converting to Long before multiplying.

PulsarHealthCheck: client.bookies() returns a handle object without making any
network call, so the check always returned healthy regardless of connectivity.
Fixed by calling client.clusters().getClusters() which makes an actual REST
call to the Pulsar admin API.

+++ cohort-api/src/main/kotlin/com/sksamuel/cohort/memory/FreememHealthCheck.kt
@@ -10,11 +10,11 @@ import java.lang.management.MemoryPoolMXBean
  *
  * The check is considered healthy if the amount of free memory is above [minFreeBytes].
  */
-class FreememHealthCheck(private val minFreeBytes: Int) : HealthCheck {
+class FreememHealthCheck(private val minFreeBytes: Long) : HealthCheck {
 
   companion object {
-    fun mb(mb: Int) = FreememHealthCheck(mb * 1024 * 1024)
-    fun gb(gb: Int) = FreememHealthCheck(gb * 1024 * 1024 * 1024)
+    fun mb(mb: Int) = FreememHealthCheck(mb.toLong() * 1024L * 1024L)
+    fun gb(gb: Int) = FreememHealthCheck(gb.toLong() * 1024L * 1024L * 1024L)
   }
 
   override val name: String = "free_mem"
